+++ src/battle_anim_utility_funcs.c
@@ -365,7 +365,7 @@ static void AnimTask_DrawFallingWhiteLinesOnAttacker_Step(u8 taskId)
 
 void InitStatsChangeAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
 
     sAnimStatsChangeData = AllocZeroed(sizeof(struct AnimStatsChangeData));
     for (i = 0; i < ARRAY_COUNT(sAnimStatsChangeData->data); i++)
@@ -605,7 +605,7 @@ void AnimTask_Flash(u8 taskId)
 
 static void AnimTask_Flash_Step(u8 taskId)
 {
-    u16 i;
+    u32 i;
     struct Task *task = &gTasks[taskId];
 
     switch (task->data[0])
@@ -645,7 +645,7 @@ static void AnimTask_Flash_Step(u8 taskId)
 
 static void SetPalettesToColor(u32 selectedPalettes, u16 color)
 {
-    u16 i;
+    u32 i;
 
     for (i = 0; i < 32; i++)
     {
@@ -667,7 +667,7 @@ static void SetPalettesToColor(u32 selectedPalettes, u16 color)
 void AnimTask_BlendNonAttackerPalettes(u8 taskId)
 {
     u32 battler;
-    int j;
+    u32 j;
     u32 selectedPalettes = 0;
 
     for (battler = 0; battler < MAX_BATTLERS_COUNT; battler++)
