+++ src/battle_factory_screen.c
@@ -1291,7 +1291,7 @@ static void CB2_InitSelectScreen(void)
 
 static void Select_InitMonsData(void)
 {
-    u8 i;
+    u32 i;
 
     if (sFactorySelectScreen != NULL)
         return;
@@ -1311,7 +1311,7 @@ static void Select_InitMonsData(void)
 
 static void Select_InitAllSprites(void)
 {
-    u8 i, cursorPos;
+    u32 i, cursorPos;
     s16 x;
 
     for (i = 0; i < SELECTABLE_MONS_COUNT; i++)
@@ -1337,7 +1337,7 @@ static void Select_InitAllSprites(void)
 
 static void Select_DestroyAllSprites(void)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < SELECTABLE_MONS_COUNT; i++)
         DestroySprite(&gSprites[sFactorySelectScreen->mons[i].ballSpriteId]);
@@ -1458,7 +1458,7 @@ static void Select_SetBallSpritePaletteNum(u8 id)
 
 static void Select_Task_OpenSummaryScreen(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 currMonId;
 
     switch (gTasks[taskId].tState)
@@ -1773,7 +1773,7 @@ static void CreateFrontierFactorySelectableMons(u8 firstMonId)
 
 static void CreateSlateportTentSelectableMons(u8 firstMonId)
 {
-    u8 i;
+    u32 i;
     u8 ivs = 0;
     u8 level = TENT_MIN_LEVEL;
     u32 otId = 0;
@@ -1791,7 +1791,7 @@ static void CreateSlateportTentSelectableMons(u8 firstMonId)
 
 static void Select_CopyMonsToPlayerParty(void)
 {
-    u8 i, j;
+    u32 i, j;
 
     for (i = 0; i < FRONTIER_PARTY_SIZE; i++)
     {
@@ -2024,7 +2024,7 @@ static void Select_ReshowMonSprite(void)
 
 static void Select_CreateChosenMonsSprites(void)
 {
-    u8 i, j;
+    u32 i, j;
 
     for (i = 0; i < FRONTIER_PARTY_SIZE; i++)
     {
@@ -2224,7 +2224,7 @@ static void Select_SetWinRegs(s16 mWin0H, s16 nWin0H, s16 mWin0V, s16 nWin0V)
 
 static bool32 Select_AreSpeciesValid(u16 monId)
 {
-    u8 i, j;
+    u32 i, j;
     u32 species = gFacilityTrainerMons[monId].species;
     u8 selectState = sFactorySelectScreen->selectingMonsState;
 
@@ -2849,7 +2849,7 @@ static void Swap_Task_SlideCycleBalls(u8 taskId)
 
 static void Swap_Task_SlideButtonOnOffScreen(u8 taskId)
 {
-    u8 i, j;
+    u32 i, j;
     s32 posX = 0;
     s8 deltaX = gTasks[taskId].tXIncrement;
     bool8 sliding;
@@ -3165,7 +3165,7 @@ static void Swap_Task_ScreenInfoTransitionIn(u8 taskId)
 // For switching between the swap screens with the player's / enemy's parties
 static void Swap_Task_SwitchPartyScreen(u8 taskId)
 {
-    u8 i;
+    u32 i;
     if (sFactorySwapScreen->monPicAnimating == TRUE)
         return;
 
@@ -3406,7 +3406,7 @@ static void CB2_InitSwapScreen(void)
 
 static void Swap_InitAllSprites(void)
 {
-    u8 i;
+    u32 i;
     u8 x;
     struct SpriteTemplate spriteTemplate;
 
@@ -3500,7 +3500,7 @@ static void Swap_InitAllSprites(void)
 
 static void Swap_DestroyAllSprites(void)
 {
-    u8 i, j;
+    u32 i, j;
 
     for (i = 0; i < FRONTIER_PARTY_SIZE; i++)
         DestroySprite(&gSprites[sFactorySwapScreen->ballSpriteIds[i]]);
@@ -3631,7 +3631,7 @@ static void Swap_UpdateMenuCursorPosition(s8 direction)
 
 static void Swap_HighlightActionButton(u8 actionId)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < ARRAY_COUNT(sFactorySwapScreen->pkmnForSwapButtonSpriteIds[0]); i++)
     {
@@ -3658,7 +3658,7 @@ static void Swap_HighlightActionButton(u8 actionId)
 
 static void Swap_HideActionButtonHighlights(void)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < ARRAY_COUNT(sFactorySwapScreen->pkmnForSwapButtonSpriteIds[0]); i++)
     {
@@ -4107,7 +4107,7 @@ static void Swap_TaskCantHaveSameMons(u8 taskId)
 
 static bool8 Swap_AlreadyHasSameSpecies(u8 monId)
 {
-    u8 i;
+    u32 i;
     u16 species = GetMonData(&gEnemyParty[monId], MON_DATA_SPECIES, NULL);
 
     for (i = 0; i < FRONTIER_PARTY_SIZE; i++)
